fix: Fix graphviz diagram (#68)
