+++ src/astx/variables.py
@@ -2,7 +2,7 @@
 
 from __future__ import annotations
 
-from typing import Any, Optional, cast
+from typing import Optional
 
 from public import public
 
@@ -169,66 +169,4 @@ class Variable(DataTypeOps):
         """Return a string that represents the object."""
         key = f"Variable[{self.name}]"
         value = self.name
-        return self._prepare_struct(key, value, simplified)
-
-
-@public
-class Argument(Variable):
-    """AST class for argument definition."""
-
-    mutability: MutabilityKind
-    name: str
-    type_: ExprType
-    default: Expr
-
-    def __init__(
-        self,
-        name: str,
-        type_: ExprType,
-        mutability: MutabilityKind = MutabilityKind.constant,
-        default: Expr = UNDEFINED,
-        loc: SourceLocation = NO_SOURCE_LOCATION,
-        parent: Optional[ASTNodes] = None,
-    ) -> None:
-        """Initialize the VarExprAST instance."""
-        super().__init__(name=name, loc=loc, parent=parent)
-        self.mutability = mutability
-        self.type_ = type_
-        self.default = default
-        self.kind = ASTKind.ArgumentKind
-
-    def __str__(self) -> str:
-        """Return a string that represents the object."""
-        type_ = self.type_.__name__
-        return f"Argument[{self.name}, {type_}]"
-
-    def get_struct(self, simplified: bool = False) -> ReprStruct:
-        """Return a string that represents the object."""
-        key = f"Argument[{self.name}, {self.type_}] = {self.default}"
-        value = cast(ReprStruct, self.default)
-        return self._prepare_struct(key, value, simplified)
-
-
-@public
-class Arguments(ASTNodes):
-    """AST class for argument definition."""
-
-    def __init__(self, *args: Argument, **kwargs: Any) -> None:
-        super().__init__(**kwargs)
-        for arg in args:
-            self.append(arg)
-
-    def __str__(self) -> str:
-        """Return a string that represents the object."""
-        return f"Arguments({len(self.nodes)})"
-
-    def get_struct(self, simplified: bool = False) -> ReprStruct:
-        """Return a string that represents the object."""
-        args_nodes = []
-
-        for node in self.nodes:
-            args_nodes.append(node.get_struct(simplified))
-
-        key = str(self)
-        value = cast(ReprStruct, args_nodes)
         return self._prepare_struct(key, value, simplified)
