+++ tests/examples/__init__.py
@@ -1,0 +1,1 @@
+"""Test with real examples."""
